# plot_wireframe plotting speedup