--- a/lib/matplotlib/transforms.py
+++ b/lib/matplotlib/transforms.py
@@ -46,8 +46,7 @@ import numpy as np
 from numpy.linalg import inv
 
 from matplotlib import _api
-from matplotlib._path import (
-    affine_transform, count_bboxes_overlapping_bbox, update_path_extents)
+from matplotlib._path import affine_transform, count_bboxes_overlapping_bbox
 from .path import Path
 
 DEBUG = False
@@ -868,8 +867,8 @@ class Bbox(BboxBase):
         if path.vertices.size == 0:
             return
 
-        points, minpos, changed = update_path_extents(
-            path, None, self._points, self._minpos, ignore)
+        points, minpos, changed = self._calc_extents_from_path(path, ignore,
+                                                               updatex, updatey)
 
         if changed:
             self.invalidate()
@@ -879,6 +878,60 @@ class Bbox(BboxBase):
             if updatey:
                 self._points[:, 1] = points[:, 1]
                 self._minpos[1] = minpos[1]
+
+    def _calc_extents_from_path(self, path, ignore, updatex=True, updatey=True):
+        """
+        Calculate the new bounds and minimum positive values for a `Bbox` from
+        the path.
+
+        Parameters
+        ----------
+        path : `~matplotlib.path.Path`
+        ignore : bool
+            - When ``True``, ignore the existing bounds of the `Bbox`.
+            - When ``False``, include the existing bounds of the `Bbox`.
+        updatex : bool
+            When ``True``, update the x-values.
+        updatey : bool
+            When ``True``, update the y-values.
+
+        Returns
+        -------
+        points : (2, 2) array
+        minpos : (2,) array
+        changed : bool
+        """
+        if ignore:
+            points = np.array([[np.inf, np.inf], [-np.inf, -np.inf]])
+            minpos = np.array([np.inf, np.inf])
+        else:
+            points = self._points.copy()
+            minpos = self._minpos.copy()
+
+        if updatex and updatey:
+            where = (np.isfinite(path.vertices[..., 0])
+                     & np.isfinite(path.vertices[..., 1]))
+        elif updatex:
+            where = np.isfinite(path.vertices[..., 0])
+        elif updatey:
+            where = np.isfinite(path.vertices[..., 1])
+        else:
+            return points, minpos, False
+
+        if updatex:
+            x = path.vertices[..., 0][where]
+            points[0, 0] = min(points[0, 0], np.min(x, initial=np.inf))
+            points[1, 0] = max(points[1, 0], np.max(x, initial=-np.inf))
+            minpos[0] = min(minpos[0], np.min(x[x > 0], initial=np.inf))
+        if updatey:
+            y = path.vertices[..., 1][where]
+            points[0, 1] = min(points[0, 1], np.min(y, initial=np.inf))
+            points[1, 1] = max(points[1, 1], np.max(y, initial=-np.inf))
+            minpos[1] = min(minpos[1], np.min(y[y > 0], initial=np.inf))
+
+        changed = np.any(points != self._points) or np.any(minpos != self._minpos)
+
+        return points, minpos, changed
 
     def update_from_data_x(self, x, ignore=None):
         """