… lib/matplotlib/_path.pyi